+++ database/migrations/2024_06_02_193004_create_extra_public_events_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('photo')->nullable();
             $table->string('name');
             $table->string('address');
-            $table->float('ticket_price');
+            $table->decimal('ticket_price', 15);
             $table->integer('reserved_tickets')->default(0);
             $table->foreignId('reservation_id')->constrained()->cascadeOnDelete();
             $table->timestamps();
